bug fixed : generIds is greater than 10000 so cannot covert error fixed in cache

+++ app/src/main/java/com/ptut/ptmovie/feature/moviedetail/MovieDetailVO.kt
@@ -8,7 +8,7 @@ import javax.inject.Inject
 data class MovieDetailVO(
     val adult: Boolean,
     val backdropPath: String,
-    val genreIds: List<Int>,
+    val genreIds: List<Long>,
     val id: Long,
     val originalLanguage: String,
     val originalTitle: String,
